services/ai-agent/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1
google-genai>=1.0.0
